+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- `v0.0.9` Badge component & `v0.0.9` Chip component
+  - Add HTML title attribute holding full label
+
+- `v0.0.6` List component
+  - Always send selected options to data provider
+  - Always send query to data provider to avoid having to extract it from filters.
+
+- `v0.0.2` ListFilterOptionItem component
+  - Set `aria-label` and `title` based on option's label
+  - Add support for long-label options (clamp to 3 lines)
+
 ## 1.8.25 - 2025-06-09
 ### Fixed
 - `v0.0.30` Form component